Added `[data-faster-link-refresh-excluded]` key related to `[data-faster-link]` to prevent FasterJs core system from refreshing this element in anytime using `refresh()` tool.
This helps to avoid an unexpected errors of complex events related to each element (for example: triggering Bootstrap Tooltip events with continuous refreshing FasterJs navigating links).

By default, this method `refresh()` is being triggered in everytime that the user calls `FasterJs.tools.dom.append||prepend||before||after` methods in a module-routes mode.

In addition when calling `generateLinks()` tool included in `refresh()` method, FasterJs replaces `[data-faster-link]` with the final actual parsed link. Now, it keeps [data-faster-link] as it is and puts the parsed link in `[data-faster-link-parsed]`.